requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.8.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
dj-database-url==2.2.0
Django==5.1
django-dynamic-fixture==4.0.1
django-extensions==3.2.3
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
filelock==3.15.4
identify==2.6.0
inflection==0.5.1
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
virtualenv==20.26.3
